Add files via upload
Solving Q1 and Q2 helps in the understanding of the NN structure by finding partial derivative for different activation function.
In Q3 I set up a DNN model for Boston Housing Data set with Relu as the activation function and linear as activation function in the last layer since the data is continuous. I also plotted the predicted vs the original response variable and found a linear pattern with no outliers.