Add files via upload
In Q1 I set up a CNN model for MNIST FASHION Dataset to categorize different photos of fashion clothing into the correct categories with accuracy >85%. I also used GRADCAM to look at #1, #45, #1001, #1005, and #5666 to see which part of the images have been scrutinized to classify it into the respective category.

In Q2 I worked with diabetes.R dataset to fit 3 types of NN models - Variational Nueral Network, NN with MC dropout and NN with quantile regression. These helps us to understand the UQ(Uncertainty Quantification) of the model.